# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Question": {
				"properties": {
					"id": {
						"type": "string"
					},
					"exam": {
						"$ref": "#/components/schemas/Exam"
					},
					"subject": {
						"type": "string",
						"description": "Enunciado da _Questão_"
					},
					"description": {
						"type": "string",
						"description": "Descrição da _Questão_"
					},
					"type": {
						"$ref": "#/components/schemas/QUESTION_TYPE",
						"description": "Tipos de _Questão_"
					},
					"alternatives": {
						"items": {
							"$ref": "#/components/schemas/Alternative"
						},
						"type": "array",
						"description": "_Alternativas da Questão_"
					}
				},
				"required": [
					"id",
					"subject",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Exam": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string",
						"description": "Título"
					},
					"courseDisciplineCode": {
						"type": "string",
						"description": "Área de estudo"
					},
					"questions": {
						"items": {
							"$ref": "#/components/schemas/Question"
						},
						"type": "array",
						"description": "_Questões_"
					}
				},
				"required": [
					"id",
					"title",
					"courseDisciplineCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"QUESTION_TYPE": {
				"enum": [
					"MULTIPLE_SINGLE",
					"MULTIPLE_MANY",
					"DISCURSIVE",
					"TEXT"
				],
				"type": "string"
			},
			"Alternative": {
				"properties": {
					"id": {
						"type": "string"
					},
					"question": {
						"$ref": "#/components/schemas/Question"
					},
					"subject": {
						"type": "string",
						"description": "Enunciado da _Questão_"
					},
					"isCorrect": {
						"type": "boolean",
						"description": "Alternativa é correta ?"
					}
				},
				"required": [
					"id",
					"subject",
					"isCorrect"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateQuestionDTO": {
				"properties": {
					"subject": {
						"type": "string",
						"description": "Título"
					},
					"description": {
						"type": "string",
						"description": "Código da Disciplina"
					},
					"type": {
						"$ref": "#/components/schemas/QUESTION_TYPE"
					}
				},
				"required": [
					"subject",
					"type",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateExamDTO": {
				"properties": {
					"title": {
						"type": "string",
						"description": "Título"
					},
					"courseDisciplineCode": {
						"type": "string",
						"description": "Código da Disciplina"
					},
					"questions": {
						"items": {
							"$ref": "#/components/schemas/ICreateQuestionDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"courseDisciplineCode"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "pratiquei-exame",
		"version": "1.0.0",
		"description": "Pratique exames de múltipla escolha.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/questions/all": {
			"get": {
				"operationId": "GetAllQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Question"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Question"
				],
				"security": [],
				"parameters": []
			}
		},
		"/questions/{id}": {
			"get": {
				"operationId": "GetQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Question"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Question"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/questions/{id}/full": {
			"get": {
				"operationId": "GetQuestionFull",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Question"
										},
										{
											"items": {
												"type": "string"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Question"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/exams": {
			"post": {
				"operationId": "PostExam",
				"responses": {
					"201": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Exam"
										},
										{
											"items": {
												"type": "string"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Exam"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateExamDTO"
							}
						}
					}
				}
			}
		},
		"/exams/all": {
			"get": {
				"operationId": "GetAllExam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Exam"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Exam"
				],
				"security": [],
				"parameters": []
			}
		},
		"/exams/{id}": {
			"get": {
				"operationId": "GetExam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Exam"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Exam"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/exams/{id}/full": {
			"get": {
				"operationId": "GetExamFull",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Exam"
										},
										{
											"items": {
												"type": "string"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Exam"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}